Move all roundtrip tests from tests/unit_tests.rs to tests/roundtrip.rs
Moved tests:
- test_writer                                   -> with_trim
- test_writer_borrow                            -> with_trim_ref
- test_writer_indent                            -> with_indent
- test_writer_indent_cdata                      -> with_indent_cdata
- test_write_empty_element_attrs                -> events::empty
- test_write_attrs                              -> partial_rewrite
- test_read_write_roundtrip                     -> simple
- test_read_write_roundtrip_escape_text         -> reescape_text
